# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /gaia/resourcebinding/gaia-merged/{rbName}:
    post:
      consumes:
      - application/json
      parameters:
      - description: RBName
        in: path
        name: rbName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 选择某个聚合后的调度方案 rb
      tags:
      - ResourceBinding
...
swagger: "2.0"